fix stupid bug

+++ dataloader.py
@@ -37,7 +37,7 @@ class ImagesDS(D.Dataset):
                 save_path = os.path.join(self.img_dir, self.mode, f'{code}_s{site}.jpeg')
                 im = cv2.imread(save_path)
                 # im = cv2.resize(im, self.target_shape[-1])
-            ims.append(torch.from_numpy(im))
+            ims.append(torch.from_numpy(im, dtype=torch.float32))
         im = torch.cat(ims, dim=-1)
         im = im.permute(2, 1, 0)
         return im / 255. - 0.5
